@@ -16,15 +16,15 @@
         "issues": "https://github.com/swoole/library"
     },
     "require": {
-        "php": ">=8.0",
-        "ext-swoole": ">=5.0"
+        "php": ">=8.1",
+        "ext-swoole": ">=5.1"
     },
     "require-dev": {
         "ext-sockets": "*",
         "ext-json": "*",
         "ext-redis": "*",
         "ext-curl": "*",
-        "phpunit/phpunit": "~9.0 || ~10.0",
+        "phpunit/phpunit": "~10.0 || ~11.0",
         "swoole/ide-helper": "dev-master"
     },
     "suggest": {
