@@ -11,15 +11,16 @@ declare(strict_types=1);
 
 namespace Swoole\Coroutine;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
- * @coversNothing
  */
+#[CoversClass(WaitGroup::class)]
 class WaitGroupTest extends TestCase
 {
-    public function testWait()
+    public function testWait(): void
     {
         run(function () {
             $wg = new WaitGroup(4);
