@@ -11,13 +11,14 @@ declare(strict_types=1);
 
 namespace Swoole\FastCGI\Record;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Swoole\FastCGI;
 
 /**
  * @internal
- * @coversNothing
  */
+#[CoversClass(GetValues::class)]
 class GetValuesTest extends TestCase
 {
     protected static string $rawMessage = '01090001001107000f00464347495f4d5058535f434f4e4e5300000000000000';
