@@ -11,30 +11,25 @@ declare(strict_types=1);
 
 namespace Swoole;
 
+use PHPUnit\Framework\Attributes\CoversFunction;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
  * @coversNothing
  */
+#[CoversFunction('swoole_library_get_options')]
+#[CoversFunction('swoole_library_set_options')]
 class FunctionTest extends TestCase
 {
-    /**
-     * @covers ::swoole_library_get_options
-     * @covers ::swoole_library_set_options
-     */
-    public function testOptions()
+    public function testOptions(): void
     {
         $options = [__METHOD__ => uniqid()];
         swoole_library_set_options($options);
         $this->assertEquals($options, swoole_library_get_options());
     }
 
-    /**
-     * @covers ::swoole_library_get_option
-     * @covers ::swoole_library_set_option
-     */
-    public function testOption()
+    public function testOption(): void
     {
         $option = uniqid();
         swoole_library_set_option(__METHOD__, $option);
