@@ -11,77 +11,54 @@ declare(strict_types=1);
 
 namespace Swoole;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @internal
- * @coversNothing
  */
+#[CoversClass(MultibyteStringObject::class)]
 class MultibyteStringObjectTest extends TestCase
 {
-    /**
-     * @covers \Swoole\MultibyteStringObject::length()
-     */
-    public function testLength()
+    public function testLength(): void
     {
         $str    = 'hello world';
         $length = swoole_mbstring($str)->length();
         $this->assertEquals(strlen($str), $length);
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::indexOf()
-     */
-    public function testIndexOf()
+    public function testIndexOf(): void
     {
         $this->assertEquals(swoole_mbstring('hello swoole and hello world')->indexOf('swoole'), 6);
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::lastIndexOf()
-     */
-    public function testLastIndexOf()
+    public function testLastIndexOf(): void
     {
         $this->assertEquals(swoole_mbstring('hello swoole and hello world')->lastIndexOf('hello'), 17);
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::pos()
-     */
-    public function testPos()
+    public function testPos(): void
     {
         $this->assertEquals(swoole_mbstring('hello swoole and hello world')->pos('and'), 13);
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::rpos()
-     */
-    public function testRPos()
+    public function testRPos(): void
     {
         $this->assertEquals(swoole_mbstring('hello swoole and hello world')->rpos('hello'), 17);
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::ipos()
-     */
-    public function testIPos()
+    public function testIPos(): void
     {
         $this->assertEquals(swoole_mbstring('hello swoole AND hello world')->ipos('and'), 13);
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::substr()
-     */
-    public function testSubstr()
+    public function testSubstr(): void
     {
         $this->assertEquals(swoole_mbstring('hello swoole and hello world')
             ->substr(4, 8)->toString(), 'o swoole');
     }
 
-    /**
-     * @covers \Swoole\MultibyteStringObject::chunk()
-     */
-    public function chunk()
+    public function chunk(): void
     {
         $r            = swoole_mbstring('hello swoole and hello world')->chunk(5)->toArray();
         $expectResult = [
